SphereWorld2.

+++ Chapter4-Objects/Objects.cpp
@@ -151,22 +151,22 @@ void SpecialKeys(int key, int x, int y)
 {
 	if (key == GLUT_KEY_UP)
 	{
-		objectFrame.RotateWorld(m3dDegToRad(-5.0f), 1.0f, 0.0f, 0.0f);
+		objectFrame.RotateWorld((float)m3dDegToRad(-5.0f), 1.0f, 0.0f, 0.0f);
 	}
 
 	if (key == GLUT_KEY_DOWN)
 	{
-		objectFrame.RotateWorld(m3dDegToRad(5.0f), 1.0f, 0.0f, 0.0f);
+		objectFrame.RotateWorld((float)m3dDegToRad(5.0f), 1.0f, 0.0f, 0.0f);
 	}
 
 	if (key == GLUT_KEY_LEFT)
 	{
-		objectFrame.RotateWorld(m3dDegToRad(-5.0f), 0.0f, 1.0f, 0.0f);
+		objectFrame.RotateWorld((float)m3dDegToRad(-5.0f), 0.0f, 1.0f, 0.0f);
 	}
 
 	if (key == GLUT_KEY_RIGHT)
 	{
-		objectFrame.RotateWorld(m3dDegToRad(5.0f), 0.0f, 1.0f, 0.0f);
+		objectFrame.RotateWorld((float)m3dDegToRad(5.0f), 0.0f, 1.0f, 0.0f);
 	}
 
 	glutPostRedisplay();
